# Conflicts:
#	app/src/main/java/com/lptiyu/tanke/activities/imagedistinguish/ImageDistinguishActivity.java
第二个版本，游戏界面，定向乐跑

+++ app/src/main/java/com/lptiyu/tanke/enums/GameType.java
@@ -4,7 +4,7 @@ package com.lptiyu.tanke.enums;
  * Created by Jason on 2016/7/18.
  */
 public interface GameType {
-    long TEAM_TYPE = 1;
-    long INDIVIDUAL_TYPE = 2;
+    int TEAM_TYPE = 1;
+    int INDIVIDUAL_TYPE = 2;
 
 }
